Create Minimum and Maximum number from inputs
In this problem, you would ask the user to enter some integers. User can enter as
many numbers as he wants to. When the user wants to stop entering the
numbers, he would press 0. When the user presses 0, you would display the
maximum number and minimum number entered by the User.